+++ openbank-vop-service/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [0.5.0](https://github.com/JiRaska/open-bank-oss/compare/vop-service-v0.4.1...vop-service-v0.5.0) (2026-08-24)
+
+
+### Features
+
+* **testing:** enforce synthetic taint REST boundaries ([#6724](https://github.com/JiRaska/open-bank-oss/issues/6724)) ([569c856](https://github.com/JiRaska/open-bank-oss/commit/569c85624aa3d6f1933865ae7ebcb69589d7d60d))
+
 ## [0.4.1](https://github.com/JiRaska/open-bank-oss/compare/vop-service-v0.4.0...vop-service-v0.4.1) (2026-08-22)